mnist-knn/requirements.txt
numpy
scikit-learn
matplotlib
pillow
joblib
keras
tensorflow
jupyterlab
